Restoring WA object in the iframe_api.ts file to avoid breaking compatibility
The WA object was missing in window otherwise because window.d.ts was not bundled in the package.
