+++ front/package.json
@@ -17,7 +17,7 @@
     "eslint": "^8.4.1",
     "eslint-plugin-svelte3": "^3.2.1",
     "jasmine": "^3.5.0",
-    "lint-staged": "^11.0.0",
+    "lint-staged": "^12.3.7",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
     "prettier-plugin-svelte": "^2.5.0",
